github co piolet/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
httpx==0.25.1
yt-dlp==2023.10.13
ffmpeg-python==0.2.0
faster-whisper==0.9.0
transformers==4.35.0
huggingface-hub==0.19.4
python-jose==3.3.0
pytesseract==0.3.10
easyocr==1.7.1
fpdf2==2.7.6
python-i18n==0.3.9
celery==5.3.4
redis==5.0.1
pillow==10.1.0
numpy==1.26.1
pandas==2.1.2
matplotlib==3.8.1
wordcloud==1.9.2